ntp_control.c: A bunch of size_t related changes.
This file is a mess, but I'll move forward on size_t elsewhere.

+++ ntpd/refclock_zyfer.c
@@ -237,7 +237,7 @@ zyfer_receive(
 			return;
 	} else {
 		memcpy(pp->a_lastcode + pp->lencode, p, rbufp->recv_length);
-		pp->lencode += rbufp->recv_length;
+		pp->lencode += (int)rbufp->recv_length;
 		pp->a_lastcode[pp->lencode] = '\0';
 	}
